move constructor to new location
